requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
asyncpg
pydantic
pydantic-settings
python-multipart
